Assignment3_Distributed-sensors/requirements.txt
paho.mqtt
numpy
rdflib